python/moderngl/requirements.txt
glfw==2.5.2
imgui==1.4.1
numpy==1.22.3
PyOpenGL==3.1.6
